refactor(core): remove unnecessary HwpError::from conversion
- Simplify error handling by removing redundant .map_err(HwpError::from) calls
- 110 warnings reduced via clippy --auto-fix (useless_conversion)
- All tests passing, build successful

Files changed:
- crates/hwp-core/src/lib.rs
- crates/hwp-core/src/document/bodytext/mod.rs
- crates/hwp-core/src/document/docinfo/doc_data.rs
- crates/hwp-core/src/document/docinfo/mod.rs
- crates/hwp-core/src/viewer/html/ctrl_header/table/cells.rs
- crates/hwp-core/src/viewer/html/ctrl_header/table/render.rs
- crates/hwp-core/src/viewer/html/ctrl_header/table/svg/borders.rs
- crates/hwp-core/src/viewer/html/ctrl_header/table/svg/fills.rs
- crates/hwp-core/src/viewer/html/document.rs
- crates/hwp-core/src/viewer/html/paragraph.rs
- crates/hwp-core/src/viewer/markdown/document/bodytext/para_text.rs
- crates/hwp-core/src/viewer/markdown/document/bodytext/table.rs
- crates/hwp-core/src/viewer/markdown/renderer.rs
- crates/hwp-core/src/viewer/markdown/utils.rs

+++ crates/hwp-core/src/viewer/html/ctrl_header/table/render.rs
@@ -349,7 +349,7 @@ pub fn render_table(
                         + para
                             .control_char_positions
                             .iter()
-                            .map(|cc| ControlChar::get_size_by_code(cc.code) as usize)
+                            .map(|cc| ControlChar::get_size_by_code(cc.code))
                             .sum::<usize>();
                 }
 
@@ -519,7 +519,7 @@ pub fn render_table(
                             + para
                                 .control_char_positions
                                 .iter()
-                                .map(|cc| ControlChar::get_size_by_code(cc.code) as usize)
+                                .map(|cc| ControlChar::get_size_by_code(cc.code))
                                 .sum::<usize>();
                         continue;
                     }
@@ -533,7 +533,7 @@ pub fn render_table(
                         + para
                             .control_char_positions
                             .iter()
-                            .map(|cc| ControlChar::get_size_by_code(cc.code) as usize)
+                            .map(|cc| ControlChar::get_size_by_code(cc.code))
                             .sum::<usize>();
                 }
 
@@ -655,7 +655,7 @@ pub fn render_table(
                                             .control_char_positions
                                             .iter()
                                             .map(|cc| {
-                                                ControlChar::get_size_by_code(cc.code) as usize
+                                                ControlChar::get_size_by_code(cc.code)
                                             })
                                             .sum::<usize>(),
                                 )
@@ -670,7 +670,7 @@ pub fn render_table(
                                     + para
                                         .control_char_positions
                                         .iter()
-                                        .map(|cc| ControlChar::get_size_by_code(cc.code) as usize)
+                                        .map(|cc| ControlChar::get_size_by_code(cc.code))
                                         .sum::<usize>(),
                             )
                         };
